% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bidirectional_motifs3}
\alias{bidirectional_motifs3}
\title{Table of motifs}
\format{
a matrix of 64 rows (1 row per motif) and 13 columns:
\itemize{
\item \code{mid}: motif identifier
\item \code{i_to_j}: a logical, is there an interaction from \code{i} to \code{j};
\item \code{j_to_i}: a logical, is there an interaction from \code{j} to \code{i};
\item \code{i_to_k}: a logical, is there an interaction from \code{i} to \code{k};
\item \code{k_to_i}: a logical, is there an interaction from \code{k} to \code{i};
\item \code{j_to_k}: a logical, is there an interaction from \code{j} to \code{k};
\item \code{k_to_j}: a logical, is there an interaction from \code{k} to \code{j};
\item \code{pid_i}: position identifier of \code{i};
\item \code{pid_j}: position identifier of \code{j};
\item \code{pid_k}: position identifier of \code{k};
\item \code{psum}: sum of positions (could be used to identify motifs);
\item \code{motif}: motifs description using ascii characters;
\item \code{name_uni}: motif name (in unidirectionnal networks).
}
}
\usage{
bidirectional_motifs3
}
\description{
Display all the 64 bidirectional 3-species motifs, for every motif,  details interactions and name notable motifs.
}
\references{
\itemize{
\item Milo, R. (2002). Network Motifs: Simple Building Blocks of Complex Networks. Science, 298(5594), 824–827. https://doi.org/10.1126/science.298.5594.824
}
}
\keyword{datasets}
