% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{motif_census_bidirectional}
\alias{motif_census_bidirectional}
\title{Compute motif census for a bidirectional network.}
\usage{
motif_census_bidirectional(mat)
}
\arguments{
\item{mat}{a square matrix of logical describing a directional network.}
}
\description{
Count motifs and positions within motifs for a bidirectional network.
}
\references{
Milo, R. (2002). Network Motifs: Simple Building Blocks of Complex Networks. Science, 298(5594), 824–827. https://doi.org/10.1126/science.298.5594.824
}
