% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table64}
\alias{table64}
\title{Table of motifs}
\format{
a matrix of 64 rows and 13 columns
}
\usage{
table64
}
\description{
Display all the 64 bidirectional motifs, details interactions and name notable motifs.
}
\keyword{datasets}
