% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{motif_census}
\alias{motif_census}
\title{Compute motif census}
\arguments{
\item{mat}{a matrix of logical describing a network.}
}
\value{
A dataframe with species as row and motifs as columns, ordered as follows:
\itemize{
\item 1 linear chains bottom
\item 2 linear chains middle
\item 3 linear chains top
\item 4 apparent compitition botton
\item 5 apparent compitition top
\item 6 expoitative competition bottom
\item 7 expoitative competition top
\item 8 omnivory bottom
\item 9 omnivory middle
\item 10 omnivory top
}
}
\description{
Compute motif census
}
