% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_census.R, R/zzz.R
\name{motif_census}
\alias{motif_census}
\title{motif_census}
\usage{
motif_census(mat, unidirectional = FALSE)
}
\arguments{
\item{mat}{a square matrix of logical describing a network.}

\item{unidirectional}{a logical. Is the network considered unidirectional.
Default is set to \code{FALSE} meaning that the network is bidirectional.}
}
\value{
A list of two elements:
\itemize{
\item motifs: motif count.
\item positions: position count.
}

If \code{unidirectional} is \code{TRUE}, then motifs ordered as follows:
\itemize{
\item 1 linear chains bottom;
\item 2 apparent competition;
\item 3 exploitative competition;
\item 4 omnivory;
\item 5 circular;
and positions are as follows:
\item 1 linear chains bottom;
\item 2 linear chains middle;
\item 3 linear chains top;
\item 4 apparent competition botton;
\item 5 apparent competition top;
\item 6 exploitative competition bottom;
\item 7 exploitative competition top;
\item 8 omnivory bottom;
\item 9 omnivory middle;
\item 10 omnivory top;
\item 11 circular (e.g. 1->2->3->1).
}

If \code{mat} corresponds to a  bidirectional network, then the 13 motifs are
ordered following Milo (2002). Positions to be detailed.
}
\description{
Compute 3 species motif census for unidirection and bidirectional
unipartite networks.
}
\references{
Milo, R. (2002). Network Motifs: Simple Building Blocks of Complex Networks. Science, 298(5594), 824–827. https://doi.org/10.1126/science.298.5594.824
}
